+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 * changed assets version to time in debug mode and moved version info to `package.json`
 * added template and js for simple image carousel
 * added lexy short renderer `bigthumbnail` to rljbase theme
-* added events `multiplane.getposts.before`, `multiplane.page` and `multiplane.getimage.before`
+* added events `multiplane.getposts.before`, `multiplane.page`, `multiplane.getimage.before`, `multiplane.sitemap`
 * added sort order to `getNav` function
 * added clearcache route (only in debug mode)
+* added sitemap
 * introduced theme config file
 * improved video field template - more php, less js
 * improved search